+++ src/src-mode.ts
@@ -1,0 +1,11 @@
+export const enum SrcMode {
+  /**
+   * The Src string is a file name.
+   */
+  File = 0,
+
+  /**
+   * The Src string is a json string.
+   */
+  Json = 1
+}